requirements.txt
requests==2.31.0
feedparser==6.0.11
python-dotenv==1.0.0
anthropic==0.25.1
html2image==2.0.4.3
Pillow==10.3.0
redis==5.0.4
